refactor: detail-dialog 상태 분리
