+++ frontend/src/pages/search/components/detail-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './detail-dialog'
